feat!: Improve static jwt creator methods
The previous methods where dubious in their naming. The same method was
overloaded to many times and placed to much meaning on the types of
arguments passed to it.

In some places we had ``fromMappedXXX () `` methods while we did not
have such methods in other places.

In addition to streamlining these public static constructor methods we
also increased the test coverage of the ``JsonWebtoken`` class.