feat: Support replying with futures from the request
Implementations where expected to accept the responses as the return
value from user defined methods. Now implementations are expected to
accept futures as the return type of user defined methods.

Those futures can be created with the ``reply ()`` method on the request
object. Implementations can provided logic in that method to properly
handle responses as they see fit, the same way as they otherwise would
have.