feat: Give default implementation to ``extractWrapped ()``
If implementors don't have any objects that they want to expose to the
outside they are still required to overwrite this method. We give a
default implementation that complains when called. It is expected that
implementors also throw a similar exception when a class is passed that
is not supported.