feat: Add common shortcut response types
We provide a way to respond to an api request by serializing an object.
Under the hood it uses jackson to do the serialization. This limits us
to always provide json responses.

We introduce three different response types:

- An empty response. No bytes are send in the response body.
- A raw string response. We send a string as is, no escaping is
  performed, and no leading or trailing double quotes are added.
- A raw bytes response. We send bytes as is, no transformations are
  performed, or encodings applied. This does only work when serializing
  to an output stream. When serializing to a string the bytes are
  provided in base64 encoded form.

If api adaptors serialize the responses to an output stream the apis
have an implementation agnostic way to respond with more than just json.