feat: Allow replying with headers of typed values
We had a type signature that required a map with values of type object,
even though strings are of type object we could not pass a map of type
string to string as an argument.

We fix the signature on the reply method of the request interface such
that the values themselves can have any type.