improve: Add return type on query decode methods
The ``decodeInto`` method did not return anything. It was meant to
populate a bag from a query string. We can however just return the bag
that we receive as an argument. This allows in some instances for more
convenient invocation of the method. As is illustrated by the refactored
``decode`` method.